StudyMate/requirements.txt
streamlit>=1.36
pymupdf>=1.24
sentence-transformers>=3.0
faiss-cpu>=1.8
numpy>=1.26
python-dotenv>=1.0
ibm-watsonx-ai>=1.1.14
